Move evaluator implemention into a private module
