+++ ert3/evaluator/__init__.py
@@ -1,1 +1,1 @@
-from .evaluator import evaluate
+from ert3.evaluator._evaluator import evaluate
